Remove long double: use nestor kronmult

+++ src/main.cpp
@@ -22,7 +22,7 @@
 #include <numeric>
 
 #ifdef ASGARD_USE_LONG_DOUBLE_PREC
-using prec = long double;
+using prec = double;
 #else
 #ifdef ASGARD_USE_DOUBLE_PREC
 using prec = double;
